pyproject.toml
[project]
name = "neon-billing-alerts"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "neon-api>=0.3.0",
    "python-dotenv>=1.2.1",
]
